--hours option and conftest.py location
